fix: 이벤트 리스트 수정 타입 반영
